feat : 재고 관리 CRUD [#57]
- 새로운 재고 추가 구현
- 재고 DELETE 구현

Resolves #57

+++ src/main/java/dev/muin/backend/domain/Product/ProductRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<Product, Short> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
     @Query("SELECT new dev.muin.backend.service.dto.StockAndProductDto(p) FROM Product p")
     List<StockAndProductDto> findAllProducts();
 }
